Moving document menu to  document list
