+++ web/src/client/js/components/Document/_Document.scss
@@ -23,11 +23,6 @@
     position: sticky;
     top: 0;
     z-index: 10;
-
-    .dropdown {
-      padding-left: 1.2rem;
-    }
-
   }
 
   &__button-expand {
