+++ web/src/client/js/components/DocumentsList/DocumentsListItem.js
@@ -5,9 +5,17 @@ import moment from 'moment'
 import cx from 'classnames'
 
 import Constants from 'Shared/constants'
-import { TimeIcon } from 'Components/Icons'
+import { MoreIcon, TimeIcon } from 'Components/Icons'
+import { DropdownItem, DropdownComponent } from 'Components/Dropdown/Dropdown'
 
-const DocumentsListItem = ({ disable, disableDragging, document, fieldCopyHandler, fieldMoveHandler }) => {
+const DocumentsListItem = ({
+  disable,
+  disableDragging,
+  document,
+  fieldCopyHandler,
+  fieldMoveHandler,
+  removeDocumentHandler,
+}) => {
   const [draggedOver, setDraggedOver] = useState(false)
 
   function dropHandler(e) {
@@ -54,6 +62,11 @@ const DocumentsListItem = ({ disable, disableDragging, document, fieldCopyHandle
     setDraggedOver(true)
   }
 
+  const dropdownButton = {
+    className: 'btn btn--blank btn--with-circular-icon',
+    icon: <MoreIcon />
+  }
+
   const documentClasses = cx({
     'documents-list__item': true,
     'documents-list__item--active': draggedOver
@@ -67,14 +80,21 @@ const DocumentsListItem = ({ disable, disableDragging, document, fieldCopyHandle
       onDragEnter={(e) => { disableDragging ? null : dragEnterHandler(e) }}
       onDragLeave={(e) => { disableDragging ? null : dragLeaveHandler(e) }}
       onDragOver={(e) => { disableDragging ? null : dragOverHandler(e) }}>
-      <NavLink to={`${Constants.PATH_PROJECT}/${document.projectId}${Constants.PATH_DOCUMENT}/${document.id}`}
+      <NavLink
+        to={`${Constants.PATH_PROJECT}/${document.projectId}${Constants.PATH_DOCUMENT}/${document.id}`}
         className="btn btn--blank documents-list__button"
-        onClick={(e) => { if (disable) { e.preventDefault() } }}>
-        <span className="documents-list__name">{ document.name }</span>
-        <time className="documents-list__date" dateTime={document.updatedAt}>
-          <TimeIcon width="13" height="13" />
-          <span>{moment(document.updatedAt).fromNow()}</span>
-        </time>
+        onClick={(e) => { if (disable) { e.preventDefault() } }}
+      >
+        <div className="documents-list__name-container">
+          <span className="documents-list__name">{ document.name }</span>
+          <time className="documents-list__date" dateTime={document.updatedAt}>
+            <TimeIcon width="13" height="13" />
+            <span>{moment(document.updatedAt).fromNow()}</span>
+          </time>
+        </div>
+        <DropdownComponent align="right" button={dropdownButton} className="documents-list__document-dropdown">
+          <DropdownItem label="Delete document..." type="button" className="btn--danger" divider onClick={() => { removeDocumentHandler(document) }} />
+        </DropdownComponent>
       </NavLink>
     </li>
   )
@@ -86,6 +106,7 @@ DocumentsListItem.propTypes = {
   document: PropTypes.object.isRequired,
   fieldCopyHandler: PropTypes.func.isRequired,
   fieldMoveHandler: PropTypes.func.isRequired,
+  removeDocumentHandler: PropTypes.func.isRequired,
 }
 
 DocumentsListItem.defaultProps = {
