+++ web/src/client/js/components/DocumentsList/_DocumentsList.scss
@@ -58,6 +58,7 @@
     margin: 0;
     padding: 0;
   }
+
   &__item {
     border-top: thin solid var(--base-border-color);
     @media (prefers-reduced-motion: no-preference) {
@@ -108,6 +109,9 @@
         padding-bottom: 0;
         padding-top: 8px;
         resize: none;
+        &::selection {
+          background-color: var(--color-yellow);
+        }
       }
       &:focus {
         background-color: var(--document-list-active-item);
@@ -115,8 +119,8 @@
     } // new
   }
   &__button {
-    align-items: flex-start;
-    flex-direction: column;
+    align-items: center;
+    flex-direction: row;
     padding: calc(var(--gutter-width) / 5) calc(var(--gutter-width) / 4);
     white-space: normal;
     width: 100%;
@@ -144,6 +148,7 @@
   &__name {
     border: 0;
     color: var(--theme-text-color);
+    display: block;
     font-weight: 400;
     padding-bottom: 6px;
     .active & {
@@ -187,4 +192,15 @@
     }
   }
 
+}
+
+
+.documents-list__name-container {
+  flex: 1;
+}
+
+.documents-list__document-dropdown {
+  display: flex;
+  justify-content: flex-end;
+  width: 48px;
 }
